requirements.txt
python-dotenv
flask-mail
flask
